backend/pyproject.toml
[project]
name = "talk2folder-backend"
version = "0.1.0"
description = "Chat with your Google Drive folders using AI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    
    # Database
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",
    
    # Authentication
    "authlib>=1.3.2",
    "httpx>=0.27.2",
    "python-jose[cryptography]>=3.3.0",
    
    # Google APIs
    "google-api-python-client>=2.154.0",
    "google-auth>=2.36.0",
    
    # Vector Store
    "chromadb>=1.4.0",
    
    # File Processing
    "pypdf>=5.1.0",
    "python-docx>=1.1.2",
    "openpyxl>=3.1.5",
    "python-pptx>=1.0.0",
    
    # AI / Gemini
    "google-generativeai>=0.8.0",
    
    # Streaming
    "sse-starlette>=2.0.0",
    
    # Utilities
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.1",
    "python-dotenv>=1.0.1",
    "tenacity>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
